Move RPC verifier/signer/tx sender to own file.
This creates a new game-channel source file, "rpcwallet".  The RPC
signature code as well as transaction sender is moved there, so that
the signatures and movesender files themselves have no traces left
of any RPC references.
