+++ gamechannel/chaintochannel_tests.cpp
@@ -13,8 +13,6 @@
 
 #include <xayautil/hash.hpp>
 
-#include <jsonrpccpp/common/exception.h>
-
 #include <google/protobuf/text_format.h>
 
 #include <gmock/gmock.h>
