+++ gamechannel/channelmanager_tests.cpp
@@ -9,8 +9,6 @@
 
 #include "proto/broadcast.pb.h"
 
-#include <jsonrpccpp/common/exception.h>
-
 #include <google/protobuf/text_format.h>
 #include <google/protobuf/util/message_differencer.h>
