+++ gamechannel/movesender_tests.cpp
@@ -8,8 +8,6 @@
 
 #include <xayautil/hash.hpp>
 
-#include <jsonrpccpp/common/exception.h>
-
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
